Moves Document into Document directory
