+++ lib/pages/Document/Document.js
@@ -1,10 +1,10 @@
 import { useRouter } from "next/router";
 import { useQuery } from "@apollo/client";
-import { Navigation } from "../components/Layout/Navigation/Navigation";
-import { Main } from "../components/Layout/Main";
-import { ArtboardList } from "../components/ArtboardList/ArtboardList";
-import { GET_DOCUMENT } from "../gql/queries/getDocumentQuery";
-import { ArtboardListItem } from "../components/ArtboardList/ArtboardListItem";
+import { Navigation } from "../../components/Layout/Navigation/Navigation";
+import { Main } from "../../components/Layout/Main";
+import { ArtboardList } from "./ArtboardList/ArtboardList";
+import { GET_DOCUMENT } from "../../gql/queries/getDocumentQuery";
+import { ArtboardListItem } from "./ArtboardList/ArtboardListItem";
 
 function Document() {
   const {
