+++ lib/pages/test/Document.test.js
@@ -1,7 +1,7 @@
 import * as nextRouter from "next/router";
 import { render, waitFor } from "@testing-library/react";
 import { MockedProvider } from "@apollo/client/testing";
-import { Document } from "../Document";
+import { Document } from "../Document/Document";
 import { GET_DOCUMENT } from "../../gql/queries/getDocumentQuery";
 import { ThemeProvider } from "styled-components";
 import { theme } from "../../styles/theme";
