+++ pages/documents/[documentShortId]/index.js
@@ -1,3 +1,3 @@
-import { Document } from "../../../lib/pages/Document";
+import { Document } from "../../../lib/pages/Document/Document";
 
 export default Document;
